+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ Change Log
     * Show title bar instead of custom title to allow users to close
       dialogs with X button
 * ``ide.scroll_to_line`` optimized
+* Toggle comments various fixes
+    * Line is properly commented when there's inline comment
 
 
 ``0.0.23`` (2017-09-14)